feat: add proto converters and visitor
This change adds some more logic to the substrait client. 1) it adds
logic to map proto messages to domain objects for relations and
expressions. 2) it adds a high level visitor control flow.
This commit also introduces a lineage extractor application to drive
development and testing of the control flow. For the testing it
leverages pre-constructed substrait plans.