feat: add a basic visitor for Substrait Relations (#2)
The main goal of this PR is to start building Substrait's c-sharp
client. The commit adds a visitor to process the SQL logical operations
represented in Substrait. The visitor's design is influenced by the
Substrait-java client. As such, this commit adds classes corresponding
to various Substrait relations in protobuf. The intention is to build
the client iteratively, and hence this commit makes a tiny improvement.

In addition, the change also updates the Substrait submodule. It also
deletes the old bindings since there is no need to track the bindings
in source control (the binding will be generated at build time).
Similarly, this commit adds a project structure which follows c-sharp
project naming conventions. Hence a new project is added, which replaces
the old core project in the solution.